select_case/racer.go
package select_case

import (
	"fmt"
	"net/http"
	"time"
)

var tenSecondTimeout = 10 * time.Second

func Racer(urlA, urlB string) (winner string, err error) {
	return ConfigurableRacer(urlA, urlB, tenSecondTimeout)
}

// ConfigurableRacer returns the winner's URL based on response time of
// urlA and urlB
func ConfigurableRacer(urlA, urlB string, timeout time.Duration) (winner string, err error) {
	timer := time.After(timeout)
	select {
	case <-ping(urlA):
		return urlA, nil
	case <-ping(urlB):
		return urlB, nil
	case <-timer:
		fmt.Println("timed out!")
		return "", fmt.Errorf("timed out waiting for %s and %s", urlA, urlB)
	}
}

func ping(url string) chan struct{} {
	ch := make(chan struct{})

	go func() {
		http.Get(url)
		close(ch)
	}()

	return ch
}
